requirements.txt
# Python 3.10+ recommended. Install: pip install -r requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-groq>=0.0.1
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
transformers>=4.46.0
huggingface_hub>=0.25.0
tavily-python>=0.3.0
python-multipart
httpx>=0.26.0
requests>=2.31.0
slowapi>=0.1.9
flask>=3.0.0
pymysql>=1.1.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1.0
python-jose[cryptography]>=3.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
murf>=1.0.0
APScheduler>=3.10.0
faster-whisper>=1.0.0
av>=10.0.0